Fix disabled spawnpoints drawing on command map
The 'team_t' enum was changed to be 'char' type in e4abfe27778adb5006bbae3e04f4d83cf4e838fd (and later to 'int8_t' in 2ff4e7ce4e451ae38086166fdd61398ca2dac9b4). This broke the functionality for hiding disabled spawns, because they have the 8th bit set to indicate being disabled. This caused a wraparound bug with the spawn parsing, where the client parsed all spawns as if the disabled bit was never set, causing all spawns to be visible at all times.

+++ src/game/bg_public.h
@@ -1528,7 +1528,7 @@ typedef struct headAnimation_s {
 inline constexpr int ANIM_TOGGLEBIT = 1 << (ANIM_BITS - 1);
 
 // Gordon: renamed these to team_axis/allies, it really was awful....
-enum team_t : int8_t {
+enum team_t : int32_t {
   TEAM_FREE,
   TEAM_AXIS,
   TEAM_ALLIES,
